feat(service): add Nonced middleware and add IsNonced
The Nonced middleware wraps a handler into NoncedHandlerFunc
requiring any request to have a valid nonce or any condition that
makes the service IsNonced false.

The IsNonced method added to the service is a way to unblock
requests from a handler or function to be allowed to proceed
without a nonce validation. This method was added to allow a
handler to have a request that will provide a new nonce.

IsNonced can be used for requests where nonce is not necessary
after a validation.

Refs: #2